% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_nb_class.R
\docType{class}
\name{abun_nb-class}
\alias{abun_nb-class}
\title{abun_nb class}
\description{
A S4 class is used to present the data and maximum penalized empirical likelihood (EL) estimates under the zero-truncated negative binomial regression model.
}
\section{Slots}{

\describe{
\item{\code{method}}{A character. \code{"EL"} and \code{"PEL"} stand for the EL and penalized EL estimation methods, respectively.}

\item{\code{N}}{A number, the maximum EL or PEL estimate of the population size.}

\item{\code{Nchao}}{A number, the Chao (1987)'s lower bound.}

\item{\code{beta}}{A vector, the maximum EL or penalized EL estimates of regression coefficients.}

\item{\code{k}}{A number, the maximum EL or penalized EL estimates of the dispersion parameter.}

\item{\code{alpha}}{A number, the maximum EL or penalized EL estimate of the probability of never being captured.}

\item{\code{loglikelihood}}{A number, the log EL or penalized log EL value.}

\item{\code{AIC}}{A number, the Akaike Information Criterion value.}

\item{\code{prob}}{A vector, the probability masses of individual covariates.}

\item{\code{nit}}{A number, the number of iterations of EM algorithms.}

\item{\code{pars}}{A matrix. Row represent the iterative values of parameters.}

\item{\code{loglikelihoods}}{A vector, the log-likelihood values in iteration.}

\item{\code{numCap}}{A vector, whose elements represent the number of times that individuals are captured.}

\item{\code{x}}{A matrix, the individual covariates without the constant one.}

\item{\code{eps}}{A number, the tolerance of threshold controling the convergence of log-likelihood in EM algorithms.}

\item{\code{maxN}}{A number, the maximum value for \eqn{N} when searching for the estimate of the population size \eqn{N}.}

\item{\code{Cp}}{A number, the penalty coefficient when the penalized EL method is used.}
}}

\references{
Chao, A. (1987).
Estimating the population size for capture--recapture data with unequal catchability.
\emph{Biometrics} \strong{43}, 783-791.
}
