% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_nb.R
\name{abun_nb}
\alias{abun_nb}
\title{Maximum penalized EL estimates for the zero-truncated negative binomial regression model}
\usage{
abun_nb(numCap, x, method = "EL", eps = 1e-05, maxN = NULL, N0 = NULL, Cp = 0)
}
\arguments{
\item{numCap}{A vector, whose elements represent the number of times that individuals are captured.}

\item{x}{A matrix, the individual covariates without the constant one.}

\item{method}{A character. \code{"EL"} and \code{"PEL"} stand for the EL and penalized EL estimation methods, respectively.}

\item{eps}{A number, the tolerance of threshold controling the convergence of log-likelihood in EM algorithms.}

\item{maxN}{A number, the maximum value for \eqn{N} when searching for the estimate of the population size \eqn{N}.}

\item{N0}{A number, at which the population size \eqn{N} is fixed when calculating the profile log-likelihood in the \code{\link{abun_nb_ci}} function. It is \code{NULL} when calculating the maximum penalized EL estimates.}

\item{Cp}{A number, the penalty coefficient when the penalized EL method is used.}
}
\value{
A \code{abun_nb} object.
}
\description{
Function to calculate the maximum penalized empirical likelihood (EL) estimates for the zero-truncated negative binomial regression model using the EM algorithm.
}
\examples{

# Fit the zero-truncated negative binomial regression model
numCap <- apply(blackbear[,1:8], 1, sum)
x <- data.frame(sex = blackbear$sex)
bear_el_nb <- abun_nb(numCap = numCap, x = x)
bear_pel_nb <- abun_nb(numCap = numCap, x = x, method = "PEL")
summary(bear_el_nb)
summary(bear_pel_nb)

# Calculate the EL ratio confidence interval of the abundance parameter
abun_nb_ci(bear_el_nb)
abun_nb_ci(bear_pel_nb)

}
