% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_nb_ci.R
\name{abun_nb_ci}
\alias{abun_nb_ci}
\title{Penalized EL ratio confidence interval for the zero-truncated negative binomial regression model}
\usage{
abun_nb_ci(object, level = 0.95, UB = 1e+09)
}
\arguments{
\item{object}{A \code{abun_nb} object.}

\item{level}{A number, the nominal level. Default is 0.95.}

\item{UB}{A number, the upper bound when searching for the upper limit of \eqn{N}. Default is \eqn{1e+09}.}
}
\value{
A vector, the penalized EL ratio confidence interval of \eqn{N}.
}
\description{
Penalized empirical likelihood (EL) ratio confidence interval of the abundance parameter \eqn{N} is calculated for the zero-truncated negative binomial regression model.
}
