% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_nb_class.R
\docType{class}
\name{abun_nb_summary-class}
\alias{abun_nb_summary-class}
\title{abun_nb_summary class}
\description{
A S4 class summarizes the maximum (penalized) empirical likelihood estimates for the zero-truncated negative binomial regression model
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{A matrix, which summarizes the estimates of regression coefficients.}

\item{\code{dispersion}}{A matrix, which summarizes the estimates of the dispersion parameter.}

\item{\code{abundance}}{A matrix, which summarizes the estimates of the population size.}
}}

