% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_nb.R
\name{loglikelihood_nb}
\alias{loglikelihood_nb}
\title{Penalized log EL for the zero-truncated negative binomial regression model}
\usage{
loglikelihood_nb(N, n, alpha, prob, numCap, z_mat, beta, k, Cp, Nchao)
}
\arguments{
\item{N}{A number, the population size.}

\item{n}{A number, the sample size.}

\item{alpha}{A number, the probability of never being captured.}

\item{prob}{A vector, the probability masses of covariates.}

\item{numCap}{A vector, whose elements represent the number of times that individuals are captured.}

\item{z_mat}{A matrix, where rows represent individuals captured and columns represent covariates.}

\item{beta}{A vector or matrix with a single column, the regression coefficients in the model \code{"Mh"}.}

\item{k}{A number, the maximum EL or penalized EL estimates of the dispersion parameter.}

\item{Cp}{A number, the penalty coefficient when the penalized EL method is used.}

\item{Nchao}{A number, the Chao (1987)'s lower bound.}
}
\value{
A number, the penalized empirical log-likelihood.
}
\description{
Function to calculate the penalized empirical log-likelihood (log EL) function for the zero-truncated negative binomial regression model.
}
\references{
Chao, A. (1987).
Estimating the population size for capture--recapture data with unequal catchability.
\emph{Biometrics} \strong{43}, 783-791.
}
