% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_nb_class.R
\name{show,abun_nb_summary-method}
\alias{show,abun_nb_summary-method}
\title{Function to show the \code{abun_nb_summary} object}
\usage{
\S4method{show}{abun_nb_summary}(object)
}
\arguments{
\item{object}{\code{abun_nb_summary} object}
}
\description{
Function to show the \code{abun_nb_summary} object
}
