% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_nb_class.R
\name{summary,abun_nb-method}
\alias{summary,abun_nb-method}
\title{Function to summarize the maximum (penalized) empirical likelihood estimates for the zero-truncated negative binomial regression model}
\usage{
\S4method{summary}{abun_nb}(object)
}
\arguments{
\item{object}{a \code{abun_nb} object}
}
\description{
Function to summarize the maximum (penalized) empirical likelihood estimates for the zero-truncated negative binomial regression model
}
